+++ tensorflow/core/common_runtime/shape_refiner.cc
@@ -19,6 +19,7 @@ limitations under the License.
 #include <unordered_set>
 #include <vector>
 
+#include "tensorflow/core/common_runtime/eval_const_tensor.h"
 #include "tensorflow/core/framework/common_shape_fns.h"
 #include "tensorflow/core/framework/node_def.pb.h"
 #include "tensorflow/core/framework/tensor.h"
@@ -407,301 +408,13 @@ Status ShapeRefiner::EvaluateConstantTensorForEdge(const Node* node,
                                                    int dst_idx, bool* evaluated,
                                                    Tensor* result) {
   *evaluated = false;
-
   const Edge* input_edge;
   TF_RETURN_IF_ERROR(node->input_edge(dst_idx, &input_edge));
-
-  // Simple case: the source node is a constant
-  const Node* src = input_edge->src();
-  if (src->IsConstant()) {
-    if (result->FromProto(src->def().attr().at("value").tensor())) {
-      *evaluated = true;
-      return Status::OK();
-    }
-  }
-
-  if (disable_constant_propagation_) {
-    return Status::OK();
-  }
-
-  bool is_constant_graph = false;
-  Graph subgraph(ops_registry_);
-  auto versions = subgraph.versions();
-  versions.set_producer(graph_def_version_);
-  subgraph.set_versions(versions);
-
-  // We identify the possibly constant subgraph to evaluate by
-  // recursively iterating backwards through the inputs to 'node'
-  // until we either 1) find an already existing input to our subgraph
-  // (filled in `const_inputs`), 2) Discover our graph is not constant,
-  // or 3) Hit a root node.
-  std::vector<std::pair<string, Tensor>> const_inputs;
-  TF_RETURN_IF_ERROR(ExtractConstantSubgraph(
-      input_edge->src(), &subgraph, &is_constant_graph, &const_inputs));
-  if (!is_constant_graph) {
-    return Status::OK();
-  }
-  const string output_tensor_name =
-      strings::StrCat(input_edge->src()->name(), ":", input_edge->src_output());
-  std::vector<Tensor> outputs;
-
-  // NOTE; we should pass in a function library runtime if we want
-  // to support constant-expression evaluation on functions.
-  Status s = graph_runner_.Run(&subgraph, nullptr /* function_library */,
-                               const_inputs, {output_tensor_name}, &outputs);
-
-  // If all kernels in the constant graph are not registered
-  // in the process, GraphRunner::Run may fail, in which case
-  // we cannot propagate constants, so this is best-effort.
-  if (s.ok()) {
-    *result = outputs[0];
-    *evaluated = true;
-
-    // We memoize (small) constants evaluated so far, so
-    // ExtractConstantSubgraph can avoid extracting the full
-    // subgraph.  As we build up large graphs, this avoids
-    // repeated computation of the early parts of a constant
-    // graph.
-    if (outputs[0].TotalBytes() <= kMaxTensorSize) {
-      const_tensor_map_[output_tensor_name] = outputs[0];
-    }
-  }
-  return Status::OK();
-}
-
-Status ShapeRefiner::TryToInferTensorOutputFromInputShapes(const Edge* edge,
-                                                           Tensor* output,
-                                                           bool* success) {
-  *success = false;
-  const Node* node = edge->src();
-  auto it = node_to_context_.find(node);
-  if (it == node_to_context_.end()) {
-    return errors::FailedPrecondition("Node does not have context.");
-  }
-  InferenceContext* c = it->second->get_context();
-
-  if (node->type_string() == "Shape") {
-    // If input shapes to the shape op are fully defined,
-    // we can infer the shape op's output tensor.
-    bool fully_defined_inputs = c->FullyDefined(c->input(0));
-    if (fully_defined_inputs) {
-      int input_rank = c->Rank(c->input(0));
-      Tensor t(node->output_type(0), TensorShape({input_rank}));
-      if (node->output_type(0) == DT_INT32) {
-        auto flat = t.flat<int>();
-        for (int i = 0; i < input_rank; i++) {
-          int64 dimension = c->Value(c->Dim(c->input(0), i));
-          if (!FastBoundsCheck(dimension, std::numeric_limits<int32>::max())) {
-            return errors::FailedPrecondition(
-                "Shape has output type int32, but dimension exceeds maximum "
-                "int32 value");
-          }
-          flat(i) = static_cast<int32>(dimension);
-        }
-      } else if (node->output_type(0) == DT_INT64) {
-        auto flat = t.flat<int64>();
-        for (int i = 0; i < input_rank; i++) {
-          flat(i) = c->Value(c->Dim(c->input(0), i));
-        }
-      } else {
-        return errors::FailedPrecondition(
-            "Shape has output type that is not int32 or int64");
-      }
-      *output = t;
-      *success = true;
-    }
-  } else if (node->type_string() == "Rank") {
-    bool rank_known = c->RankKnown(c->input(0));
-    if (rank_known) {
-      int32 input_rank = c->Rank(c->input(0));
-      Tensor t(node->output_type(0), TensorShape({}));
-      t.flat<int32>()(0) = input_rank;
-      *output = t;
-      *success = true;
-    }
-  } else if (node->type_string() == "Size") {
-    bool fully_defined_inputs = c->FullyDefined(c->input(0));
-    if (fully_defined_inputs) {
-      int32 rank = c->Rank(c->input(0));
-      Tensor t(node->output_type(0), TensorShape({}));
-      int64 size = 1;
-      for (int i = 0; i < rank; i++) {
-        size *= c->Value(c->Dim(c->input(0), i));
-      }
-      if (node->output_type(0) == DT_INT32) {
-        if (!FastBoundsCheck(size, std::numeric_limits<int32>::max())) {
-          return errors::FailedPrecondition(
-              "Size has output type int32, but size exceeds maximum int32 "
-              "value");
-        }
-        t.flat<int32>()(0) = static_cast<int32>(size);
-      } else if (node->output_type(0) == DT_INT64) {
-        t.flat<int64>()(0) = size;
-      } else {
-        return errors::FailedPrecondition(
-            "Size has output type that is not int32 or int64");
-      }
-      *output = t;
-      *success = true;
-    }
-  }
-  return Status::OK();
-}
-
-Status ShapeRefiner::ExtractConstantSubgraph(
-    Node* target_node, Graph* out_graph, bool* is_constant_graph,
-    std::vector<std::pair<string, Tensor>>* const_inputs) {
-  *is_constant_graph = false;
-  std::unordered_set<string> const_inputs_added;
-
-  if (target_node->op_def().is_stateful()) {
-    return Status::OK();
-  }
-
-  if (target_node->type_string() == "PlaceholderWithDefault") {
-    return Status::OK();
-  }
-
-  // TODO(skyewm): more of the filtering applied in input nodes below should be
-  // applied to target_node here
-
-  struct NodeAndRecursed {
-    Node* new_node = nullptr;
-    bool recursed = false;
-  };
-
-  std::map<Node*, NodeAndRecursed> old_to_new_and_recursed;
-  Node* target_node_copy = out_graph->CopyNode(target_node);
-  old_to_new_and_recursed[target_node].new_node = target_node_copy;
-  old_to_new_and_recursed[target_node].recursed = true;
-
-  // Add the target node's inputs to seed the recursion.
-  std::deque<const Edge*> edges_to_visit;
-  for (const Edge* e : target_node->in_edges()) {
-    // TODO(vrv): What do we do about control edges?  Based on our
-    // definition of a constant graph, we should be free to ignore
-    // control edges since the order in which a constant graph is
-    // executed should be the same regardless of when nodes run: we
-    // should only need to recurse down data edges.
-    if (e->IsControlEdge()) continue;
-    edges_to_visit.push_back(e);
-  }
-
-  *is_constant_graph = true;
-
-  // Iterate over the set of edges to visit (backwards).
-  while (!edges_to_visit.empty()) {
-    const Edge* current_edge = edges_to_visit.front();
-    edges_to_visit.pop_front();
-    Node* current_node = current_edge->src();
-
-    // If the node is stateful, assume the graph is not constant.
-    if (current_node->op_def().is_stateful()) {
-      *is_constant_graph = false;
-      return Status::OK();
-    }
-
-    // During construction or import from GraphConstructor, back edges may not
-    // be filled in.  Don't constant fold through merges at all for now.
-    if (IsMerge(current_node)) {
-      *is_constant_graph = false;
-      return Status::OK();
-    }
-
-    // Don't constant fold enter/exit currently either, as it's easy to end
-    // up with a partial frame.
-    if (IsEnter(current_node) || IsExit(current_node)) {
-      *is_constant_graph = false;
-      return Status::OK();
-    }
-
-    // Placeholders should never be constant folded because their outputs are
-    // fed by the user. Note that "Placeholder" nodes have no inputs so are
-    // handled below.
-    if (current_node->type_string() == "PlaceholderWithDefault") {
-      *is_constant_graph = false;
-      return Status::OK();
-    }
-
-    // If there is nothing more to recurse down, see if
-    // the generator node is a constant.
-    if (current_node->num_inputs() == 0) {
-      if (!current_node->IsConstant()) {
-        // Generator node is not a constant, so subgraph is not
-        // constant.
-        *is_constant_graph = false;
-        return Status::OK();
-      }
-    }
-
-    // Either the node is a constant, or the node is a potential
-    // intermediate node on the path from a constant.
-    //
-    // Add a copy of its node and a new edge to the new subgraph.
-
-    // Get or create the version of 'current_node' in the new graph.
-    Node* current_node_copy;
-    // This gets or creates the NodeAndRecursed entry for current_node.
-    NodeAndRecursed* node_and_recursed = &old_to_new_and_recursed[current_node];
-    if (node_and_recursed->new_node == nullptr) {
-      // First time processing this node.
-      current_node_copy = out_graph->CopyNode(current_node);
-      // Track the mapping from the original node to the new one.
-      node_and_recursed->new_node = current_node_copy;
-    } else {
-      current_node_copy = node_and_recursed->new_node;
-    }
-
-    // Add the edge to the destination node.
-    {
-      auto it = old_to_new_and_recursed.find(current_edge->dst());
-      if (it == old_to_new_and_recursed.end()) {
-        return errors::Internal(
-            "Could not find mapping from old to new copy of destination node: ",
-            current_edge->dst()->name());
-      }
-      Node* dst_copy = it->second.new_node;
-
-      out_graph->AddEdge(current_node_copy, current_edge->src_output(),
-                         dst_copy, current_edge->dst_input());
-    }
-
-    const string& output_tensor_name =
-        strings::StrCat(current_node->name(), ":", current_edge->src_output());
-
-    // Some tensor values can be inferred. For example, a shape op
-    // with input shapes fully defined can have its output tensor inferred.
-    Tensor tensor_inferred;
-    bool successfully_inferred_tensor = false;
-    TF_RETURN_IF_ERROR(TryToInferTensorOutputFromInputShapes(
-        current_edge, &tensor_inferred, &successfully_inferred_tensor));
-    if (successfully_inferred_tensor) {
-      const_inputs->emplace_back(output_tensor_name, tensor_inferred);
-      const_inputs_added.insert(output_tensor_name);
-      continue;
-    }
-
-    // If we have a copy of the input tensor materialized already,
-    // then add to the list of inputs to feed and do not recurse further.
-    auto it = const_tensor_map_.find(output_tensor_name);
-    if (it != const_tensor_map_.end() &&
-        const_inputs_added.count(output_tensor_name) == 0) {
-      const_inputs->emplace_back(output_tensor_name, it->second);
-      const_inputs_added.insert(output_tensor_name);
-      continue;
-    }
-
-    // If this node's inputs have not been processed already, do so now.
-    if (!node_and_recursed->recursed) {
-      node_and_recursed->recursed = true;
-      for (const Edge* e : current_node->in_edges()) {
-        if (e->IsControlEdge()) continue;
-        edges_to_visit.push_back(e);
-      }
-    }
-  }
-
-  return Status::OK();
+  OutputTensor tensor(input_edge->src(), input_edge->src_output());
+  return EvaluateConstantTensor(tensor, *this, *ops_registry_,
+                                graph_def_version_, evaluated, result,
+                                &graph_runner_, &const_tensor_map_,
+                                kMaxTensorSize, disable_constant_propagation_);
 }
 
 Status ShapeRefiner::ConstantPartialShape(InferenceContext* target_context,
